+++ saleor/graphql/account/types.py
@@ -11,7 +11,7 @@ from ...core.exceptions import PermissionDenied
 from ...core.permissions import AccountPermissions, AppPermission, OrderPermissions
 from ...core.tracing import traced_resolver
 from ...order import OrderStatus
-from ..account.utils import requestor_has_access
+from ..account.utils import check_requestor_access
 from ..app.dataloaders import AppByIdLoader
 from ..app.types import App
 from ..checkout.dataloaders import CheckoutByUserAndChannelLoader, CheckoutByUserLoader
@@ -170,11 +170,8 @@ class CustomerEvent(ModelObjectType):
     @staticmethod
     def resolve_app(root: models.CustomerEvent, info):
         requestor = get_user_or_app_from_context(info.context)
-        if requestor_has_access(requestor, root.user, AppPermission.MANAGE_APPS):
-            return (
-                AppByIdLoader(info.context).load(root.app_id) if root.app_id else None
-            )
-        raise PermissionDenied()
+        check_requestor_access(requestor, root.user, AppPermission.MANAGE_APPS)
+        return AppByIdLoader(info.context).load(root.app_id) if root.app_id else None
 
     @staticmethod
     def resolve_message(root: models.CustomerEvent, _info):
@@ -229,7 +226,7 @@ class User(ModelObjectType):
         description="Returns the last open checkout of this user.",
         deprecation_reason=(
             f"{DEPRECATED_IN_3X_FIELD} "
-            "Use the `checkout_tokens` field to fetch the user checkouts."
+            "Use the `checkoutTokens` field to fetch the user checkouts."
         ),
     )
     checkout_tokens = graphene.List(
@@ -364,9 +361,13 @@ class User(ModelObjectType):
         def _resolve_orders(orders):
             requester = get_user_or_app_from_context(info.context)
             if not requester.has_perm(OrderPermissions.MANAGE_ORDERS):
-                orders = list(
-                    filter(lambda order: order.status != OrderStatus.DRAFT, orders)
-                )
+                # allow fetch requestor orders (except drafts)
+                if root == info.context.user:
+                    orders = list(
+                        filter(lambda order: order.status != OrderStatus.DRAFT, orders)
+                    )
+                else:
+                    raise PermissionDenied()
 
             return create_connection_slice(
                 orders, info, kwargs, OrderCountableConnection