+++ saleor/graphql/shipping/mutations/shippings.py
@@ -17,6 +17,7 @@ from ....shipping.utils import (
     get_countries_without_shipping_zone,
 )
 from ...channel.types import ChannelContext
+from ...core.fields import JSONString
 from ...core.mutations import BaseMutation, ModelDeleteMutation, ModelMutation
 from ...core.scalars import WeightScalar
 from ...core.types.common import ShippingError
@@ -37,7 +38,7 @@ class ShippingPostalCodeRulesCreateInputRange(graphene.InputObjectType):
 
 class ShippingPriceInput(graphene.InputObjectType):
     name = graphene.String(description="Name of the shipping method.")
-    description = graphene.JSONString(description="Shipping method description.")
+    description = JSONString(description="Shipping method description.")
     minimum_order_weight = WeightScalar(
         description="Minimum order weight to use this shipping method."
     )