+++ CHANGELOG.md
@@ -8,12 +8,16 @@ All notable, unreleased changes to this project will be documented in this file.
 # Unreleased
 
 ### Breaking changes
+- Require manage orders for fetching `user.orders` - #9128 by @IKarbowiak
+  - only staff with `manage orders` and can fetch customer orders
+  - the customer can fetch his own orders, except drafts
 
 ### Other changes
 - Filter Customer/Order/Sale/Product/ProductVariant by datetime of last modification - #9137 by @rafalp
 - Add possibility for plugins to execute code before each mutation - #9193 by @NyanKiyoshi
 - Add support for hiding plugins - #9219 by @NyanKiyoshi
 - Remove `graphene-federation` dependency - #9184 by @rafalp
+- Override graphene.JSONString to have more meaningful error message #9171 by @L3str4nge
 
 
 # 3.1.0
@@ -79,6 +83,8 @@ All notable, unreleased changes to this project will be documented in this file.
 - Raise ValidationError in `CheckoutAddPromoCode`, `CheckoutPaymentCreate` when product in the checkout is
 unavailable - #8978 by @IKarbowiak
 - Remove `graphene-django` dependency - #9170 by @rafalp
+- Fix disabled warehouses appearing as valid click and collect points when checkout contains only preorders - #9052 by @rafalp
+- Fix failing `on_failure` export tasks method - #9160 by @IKarbowiak
 
 
 # 3.0.0